:label: refactor type files
