+++ src/components/bgg-search/index.ts
@@ -1,0 +1,1 @@
+export * from './bgg-search.model';
