+++ src/components/commons.models.ts
@@ -1,43 +1,1 @@
-export interface GameCard {
-  name: string;
-  editor: string;
-  year: number;
-  types: string[];
-  language: 'en' | 'es' | 'de' | 'x';
-  players: number[];
-  time: number;
-  complexity: number;
-  rate: number;
-  image: string;
-  isPlayed: boolean;
-  age: number;
-  bggReference: number;
-  size: 'xs' | 's' | 'm' | 'l';
-}
-
-export interface OracleCard {
-  name: string;
-  artist: string;
-  language: 'en' | 'es' | 'de' | 'x';
-  image?: string;
-  description: string[];
-  web: string;
-}
-
-export type SortingOrder = Record<string, boolean>;
-
-export type SortingFunctions = Record<string, (data: GameCard[]) => GameCard[]>;
-
-export interface GameDetails {
-  name: string | string[];
-  yearpublished: string;
-  minplayers: string;
-  maxplayers: string;
-  playingtime: string;
-  age: string;
-  description: string;
-  boardgamecategory: string[];
-  image: string;
-  boardgamepublisher: string[];
-  size: string[];
-}
+export * from './models';
