+++ src/components/models/bgg.model.ts
@@ -1,0 +1,13 @@
+export interface GameDetails {
+  name: string | string[];
+  yearpublished: string;
+  minplayers: string;
+  maxplayers: string;
+  playingtime: string;
+  age: string;
+  description: string;
+  boardgamecategory: string[];
+  image: string;
+  boardgamepublisher: string[];
+  size: string[];
+}
