+++ src/components/models/index.ts
@@ -1,0 +1,2 @@
+export * from './catalog.model';
+export * from './bgg.model';
